String length cannot be negative.

+++ MemoryToolKit/ProcessUtils/Read/ProcessExtensions_ReadString.cs
@@ -14,7 +14,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="ReadString(Process, int, string, int, int[])"/> overload to specify a different length.
+	///     Use the <see cref="ReadString(Process, uint, string, int, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process which contains the module.</param>
 	/// <param name="moduleName">The name of the module from which the memory should be read, including its file extension.</param>
@@ -38,7 +38,7 @@ public static partial class ProcessExtensions
 	/// <remarks>
 	///     When the <paramref name="stringType"/> parameter is any of the <c>Sized</c> <see cref="StringType"/>s, will attempt to infer the length of the string automatically.<br/>
 	///     Otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="ReadString(Process, int, StringType, string, int, int[])"/> overload to specify a different length.
+	///     Use the <see cref="ReadString(Process, uint, StringType, string, int, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process which contains the module.</param>
 	/// <param name="stringType">The type of the string in memory.</param>
@@ -69,7 +69,7 @@ public static partial class ProcessExtensions
 	///     The string read from the process' memory if the function succeeds;
 	///     otherwise, <see langword="null"/>.
 	/// </returns>
-	public static string ReadString(this Process process, int length, string moduleName, int baseOffset, params int[] offsets)
+	public static string ReadString(this Process process, uint length, string moduleName, int baseOffset, params int[] offsets)
 	{
 		if (process.Module(moduleName) is var module)
 			return process.ReadString(length, StringType.Auto, module.BaseAddress + baseOffset, offsets);
@@ -90,7 +90,7 @@ public static partial class ProcessExtensions
 	///     The string read from the process' memory if the function succeeds;
 	///     otherwise, <see langword="null"/>.
 	/// </returns>
-	public static string ReadString(this Process process, int length, StringType stringType, string moduleName, int baseOffset, params int[] offsets)
+	public static string ReadString(this Process process, uint length, StringType stringType, string moduleName, int baseOffset, params int[] offsets)
 	{
 		if (process.Module(moduleName) is var module)
 			return process.ReadString(length, stringType, module.BaseAddress + baseOffset, offsets);
@@ -105,7 +105,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="ReadString(Process, int, ProcessModule, int, int[])"/> overload to specify a different length.
+	///     Use the <see cref="ReadString(Process, uint, ProcessModule, int, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process which contains the module.</param>
 	/// <param name="module">The module from which the memory should be read.</param>
@@ -128,7 +128,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="ReadString(Process, int, StringType, ProcessModule, int, int[])"/> overload to specify a different length.
+	///     Use the <see cref="ReadString(Process, uint, StringType, ProcessModule, int, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process which contains the module.</param>
 	/// <param name="stringType">The type of the string in memory.</param>
@@ -159,7 +159,7 @@ public static partial class ProcessExtensions
 	///     The string read from the process' memory if the function succeeds;
 	///     otherwise, <see langword="null"/>.
 	/// </returns>
-	public static string ReadString(this Process process, int length, ProcessModule module, int baseOffset, params int[] offsets)
+	public static string ReadString(this Process process, uint length, ProcessModule module, int baseOffset, params int[] offsets)
 	{
 		if (module is null)
 			return null;
@@ -180,7 +180,7 @@ public static partial class ProcessExtensions
 	///     The string read from the process' memory if the function succeeds;
 	///     otherwise, <see langword="null"/>.
 	/// </returns>
-	public static string ReadString(this Process process, int length, StringType stringType, ProcessModule module, int baseOffset, params int[] offsets)
+	public static string ReadString(this Process process, uint length, StringType stringType, ProcessModule module, int baseOffset, params int[] offsets)
 	{
 		if (module is null)
 			return null;
@@ -195,7 +195,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="ReadString(Process, int, IntPtr, int[])"/> overload to specify a different length.
+	///     Use the <see cref="ReadString(Process, uint, IntPtr, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process from which the memory should be read.</param>
 	/// <param name="address">The address from which to read or start the pointer at.</param>
@@ -214,7 +214,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="ReadString(Process, int, StringType, IntPtr, int[])"/> overload to specify a different length.
+	///     Use the <see cref="ReadString(Process, uint, StringType, IntPtr, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process from which the memory should be read.</param>
 	/// <param name="stringType">The type of the string in memory.</param>
@@ -233,7 +233,7 @@ public static partial class ProcessExtensions
 
 		return stringType switch
 		{
-			StringType.AutoSized or StringType.UTF8Sized or StringType.UTF16Sized => process.ReadString(process.Read<int>(address - 0x4), stringType, address),
+			StringType.AutoSized or StringType.UTF8Sized or StringType.UTF16Sized => process.ReadString((uint)(process.Read<int>(address - 0x4)), stringType, address),
 			_ => process.ReadString(1024, stringType, address)
 		};
 	}
@@ -249,7 +249,7 @@ public static partial class ProcessExtensions
 	///     The string read from the process' memory if the function succeeds;
 	///     otherwise, <see langword="null"/>.
 	/// </returns>
-	public static string ReadString(this Process process, int length, IntPtr address, params int[] offsets)
+	public static string ReadString(this Process process, uint length, IntPtr address, params int[] offsets)
 	{
 		return process.ReadString(length, StringType.Auto, address, offsets);
 	}
@@ -266,7 +266,7 @@ public static partial class ProcessExtensions
 	///     The string read from the process' memory if the function succeeds;
 	///     otherwise, <see langword="null"/>.
 	/// </returns>
-	public static string ReadString(this Process process, int length, StringType stringType, IntPtr address, params int[] offsets)
+	public static string ReadString(this Process process, uint length, StringType stringType, IntPtr address, params int[] offsets)
 	{
 		if (length == 0)
 			return string.Empty;
@@ -295,7 +295,7 @@ public static partial class ProcessExtensions
 				_ => (Encoding.Unicode, true, 2)
 			};
 
-			length *= encoding.CharSize;
+			length *= (uint)(encoding.CharSize);
 
 			for (int j = 0; j < length; j += encoding.CharSize)
 			{
@@ -323,7 +323,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="TryReadString(Process, int, out string, string, int, int[])"/> overload to specify a different length.
+	///     Use the <see cref="TryReadString(Process, uint, out string, string, int, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process which contains the module.</param>
 	/// <param name="result">The string read from the process' memory if the function succeeds; otherwise, <see langword="null"/>.</param>
@@ -349,7 +349,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="TryReadString(Process, int, StringType, out string, string, int, int[])"/> overload to specify a different length.
+	///     Use the <see cref="TryReadString(Process, uint, StringType, out string, string, int, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process which contains the module.</param>
 	/// <param name="stringType">The type of the string in memory.</param>
@@ -384,7 +384,7 @@ public static partial class ProcessExtensions
 	///     <see langword="true"/> if the function succeeds;
 	///     otherwise, <see langword="false"/>.
 	/// </returns>
-	public static bool TryReadString(this Process process, int length, out string result, string moduleName, int baseOffset, params int[] offsets)
+	public static bool TryReadString(this Process process, uint length, out string result, string moduleName, int baseOffset, params int[] offsets)
 	{
 		result = null;
 
@@ -408,7 +408,7 @@ public static partial class ProcessExtensions
 	///     <see langword="true"/> if the function succeeds;
 	///     otherwise, <see langword="false"/>.
 	/// </returns>
-	public static bool TryReadString(this Process process, int length, StringType stringType, out string result, string moduleName, int baseOffset, params int[] offsets)
+	public static bool TryReadString(this Process process, uint length, StringType stringType, out string result, string moduleName, int baseOffset, params int[] offsets)
 	{
 		result = null;
 
@@ -425,7 +425,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="TryReadString(Process, out string, ProcessModule, int, int[])"/> overload to specify a different length.
+	///     Use the <see cref="TryReadString(Process, uint, out string, ProcessModule, int, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process which contains the module.</param>
 	/// <param name="result">The string read from the process' memory if the function succeeds; otherwise, <see langword="null"/>.</param>
@@ -451,7 +451,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="TryReadString(Process, StringType, out string, ProcessModule, int, int[])"/> overload to specify a different length.
+	///     Use the <see cref="TryReadString(Process, uint, StringType, out string, ProcessModule, int, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process which contains the module.</param>
 	/// <param name="stringType">The type of the string in memory.</param>
@@ -486,7 +486,7 @@ public static partial class ProcessExtensions
 	///     <see langword="true"/> if the function succeeds;
 	///     otherwise, <see langword="false"/>.
 	/// </returns>
-	public static bool TryReadString(this Process process, int length, out string result, ProcessModule module, int baseOffset, params int[] offsets)
+	public static bool TryReadString(this Process process, uint length, out string result, ProcessModule module, int baseOffset, params int[] offsets)
 	{
 		result = null;
 
@@ -510,7 +510,7 @@ public static partial class ProcessExtensions
 	///     <see langword="true"/> if the function succeeds;
 	///     otherwise, <see langword="false"/>.
 	/// </returns>
-	public static bool TryReadString(this Process process, int length, StringType stringType, out string result, ProcessModule module, int baseOffset, params int[] offsets)
+	public static bool TryReadString(this Process process, uint length, StringType stringType, out string result, ProcessModule module, int baseOffset, params int[] offsets)
 	{
 		result = null;
 
@@ -527,7 +527,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="TryReadString(Process, int, out string, IntPtr, int[])"/> overload to specify a different length.
+	///     Use the <see cref="TryReadString(Process, uint, out string, IntPtr, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process from which the memory should be read.</param>
 	/// <param name="result">The string read from the process' memory if the function succeeds; otherwise, <see langword="null"/>.</param>
@@ -552,7 +552,7 @@ public static partial class ProcessExtensions
 	/// </summary>
 	/// <remarks>
 	///     Unless explicitly specified otherwise, will read a maximum of 1024 characters or until a null-terminator is found.<br/>
-	///     Use the <see cref="TryReadString(Process, int, StringType, out string, IntPtr, int[])"/> overload to specify a different length.
+	///     Use the <see cref="TryReadString(Process, uint, StringType, out string, IntPtr, int[])"/> overload to specify a different length.
 	/// </remarks>
 	/// <param name="process">The process from which the memory should be read.</param>
 	/// <param name="stringType">The type of the string in memory.</param>
@@ -570,7 +570,7 @@ public static partial class ProcessExtensions
 		if ((!process?.HasExited ?? false) && process.TryDeref(out var derefAddress, address, offsets) && derefAddress != IntPtr.Zero)
 			return stringType switch
 			{
-				StringType.AutoSized or StringType.UTF8Sized or StringType.UTF16Sized => process.TryReadString(process.Read<int>(derefAddress - 0x4), stringType, out result, derefAddress),
+				StringType.AutoSized or StringType.UTF8Sized or StringType.UTF16Sized => process.TryReadString((uint)(process.Read<int>(derefAddress - 0x4)), stringType, out result, derefAddress),
 				_ => process.TryReadString(1024, stringType, out result, derefAddress)
 			};
 
@@ -589,7 +589,7 @@ public static partial class ProcessExtensions
 	///     <see langword="true"/> if the function succeeds;
 	///     otherwise, <see langword="false"/>.
 	/// </returns>
-	public static bool TryReadString(this Process process, int length, out string result, IntPtr address, params int[] offsets)
+	public static bool TryReadString(this Process process, uint length, out string result, IntPtr address, params int[] offsets)
 	{
 		result = null;
 
@@ -612,7 +612,7 @@ public static partial class ProcessExtensions
 	///     <see langword="true"/> if the function succeeds;
 	///     otherwise, <see langword="false"/>.
 	/// </returns>
-	public static bool TryReadString(this Process process, int length, StringType stringType, out string result, IntPtr address, params int[] offsets)
+	public static bool TryReadString(this Process process, uint length, StringType stringType, out string result, IntPtr address, params int[] offsets)
 	{
 		result = null;
 
@@ -650,7 +650,7 @@ public static partial class ProcessExtensions
 				_ => (Encoding.Unicode, true, 2)
 			};
 
-			length *= encoding.CharSize;
+			length *= (uint)(encoding.CharSize);
 
 			for (int j = 0; j < length; j += encoding.CharSize)
 			{
